% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{aid}
\alias{aid}
\alias{aid<-}
\title{S3 tableMatrix generic to get or set aid attribute}
\usage{
aid(obj, ...)

aid(obj) <- value
}
\arguments{
\item{obj}{Variable name}

\item{...}{Arguments passed to or from other methods.}

\item{value}{\code{data.table}.}
}
\description{
S3 tableMatrix generic to get or set aid attribute
}

