% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{copy.tableMatrix}
\alias{copy.tableMatrix}
\title{Copy}
\usage{
\method{copy}{tableMatrix}(obj)
}
\arguments{
\item{obj}{\code{tableMatrix} object.}
}
\value{
\code{tableMatrix} object.
}
\description{
Copy of a \code{tableMatrix} object.
}
