% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dataType}
\alias{dataType}
\alias{dataType<-}
\title{S3 tableMatrix generic to get or set dataType attribute}
\usage{
dataType(obj, ...)

dataType(obj) <- value
}
\arguments{
\item{obj}{Object.}

\item{...}{Passed arguments.}

\item{value}{Named list.}
}
\value{
List.
}
\description{
S3 tableMatrix generic to get or set dataType attribute
}
