% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dataType.tableList}
\alias{dataType.tableList}
\alias{dataType<-.tableList}
\alias{dataType.tableMatrix}
\alias{dataType<-.tableMatrix}
\title{Get or set dataType attribute}
\usage{
\method{dataType}{tableList}(obj)

\method{dataType}{tableList}(obj) <- value

\method{dataType}{tableMatrix}(obj)

\method{dataType}{tableMatrix}(obj) <- value
}
\arguments{
\item{obj}{\code{tableMatrix} or \code{tableList} object.}

\item{value}{List of datatypes.}
}
\value{
List of datatypes.
}
\description{
\code{tableList} and \code{tableMarix} method to get or set dataType attribute.
}
