% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dim.tableList}
\alias{dim.tableList}
\title{Dimensions}
\usage{
\method{dim}{tableList}(obj)
}
\arguments{
\item{obj}{\code{tableList} object}
}
\description{
tableList method returns dimensions of the table attribute
}

