% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dimnames.tableList}
\alias{dimnames.tableList}
\title{Dimension Names}
\usage{
\method{dimnames}{tableList}(x)
}
\arguments{
\item{x}{\code{tableList} object}
}
\description{
tableList method returns dimension names of the table attribute
}

