% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{getRow}
\alias{getRow}
\alias{setRow}
\title{S3 tableMatrix generic to get or set row from matrix attribute}
\usage{
getRow(...)

setRow(...)
}
\arguments{
\item{...}{Passed arguments.}
}
\value{
numeric vector.
}
\description{
S3 tableMatrix generic to get or set row from matrix attribute
}

