% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{getRowDim}
\alias{getRowDim}
\title{S3 tableMatrix generic to get row dim}
\usage{
getRowDim(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\description{
S3 tableMatrix generic to get row dim
}

