% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{getRowRepo}
\alias{getRowRepo}
\title{S3 tableMatrix generic to get row repo for matrix attribute}
\usage{
getRowRepo(...)
}
\arguments{
\item{...}{Passed arguments.}
}
\value{
numeric vector.
}
\description{
S3 tableMatrix generic to get row repo for matrix attribute
}
