% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\docType{data}
\name{images10By10}
\alias{images10By10}
\title{Images of distributions}
\format{A data frame with 90 rows and 103 variables:
\itemize{
  \item{direction direction of diagonal, factor "up", "down", "both"}
  \item{dimX dimension x of image, in pixels}
  \item{dimY dimension y of image, in pixels}
  \item{pixel1:pixel100 pixel of image, in number}
}}
\usage{
images10By10
}
\description{
A dataset containing samples of multivariate normal distributions. 
Each row in dataset represents a generated matrix. Dimensions are stored in
dimX and dimY. Direction "up" and "down" depends on used covariance matrix.
"Both" was created as combination of the previous.
}
\keyword{datasets}

