% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{is.tableList}
\alias{is.tableList}
\title{tableList test}
\usage{
is.tableList(obj)
}
\arguments{
\item{obj}{\code{tableList} object}
}
\value{
logical vector.
}
\description{
Tests if passed object is of class \code{tableList}
}
