% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{is.tableMatrix}
\alias{is.tableMatrix}
\title{Method to test class of object}
\usage{
is.tableMatrix(obj)
}
\arguments{
\item{obj}{\code{tableTable} object}
}
\description{
Tests if passed obj is of class tableMatrix
}

