% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{mat.tableMatrix}
\alias{mat.tableMatrix}
\alias{mat<-.tableMatrix}
\title{Get or set matrix attribute}
\usage{
\method{mat}{tableMatrix}(obj, matN = NULL)

\method{mat}{tableMatrix}(obj) <- value
}
\description{
tableMatrix method to get or set matrix attribute
}

