% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{matDim.tableMatrix}
\alias{matDim.tableMatrix}
\alias{matDim<-.tableMatrix}
\title{Get or set matDim attribute}
\usage{
\method{matDim}{tableMatrix}(obj, matN = NULL, resetN = FALSE)

\method{matDim}{tableMatrix}(obj) <- value
}
\description{
tableMatrix method to get or set matDim attribute
}

