% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{merge.tableMatrix}
\alias{merge.tableMatrix}
\title{Merging tableMatrix}
\usage{
\method{merge}{tableMatrix}(obj, dataObj, key)
}
\arguments{
\item{obj}{\code{tableMatrix} object}

\item{dataObj}{Data.table or data.frame object}

\item{key}{}
}
\description{
Merging \code{tableMatrix} with data.table or data.frame
}

