% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{print.tableList}
\alias{print.tableList}
\title{Print}
\usage{
\method{print}{tableList}(x, ...)
}
\arguments{
\item{x}{\code{tableList} object.}

\item{...}{Passed arguments.}
}
\value{
character vector
}
\description{
\code{tableList} method, prints table attribute.
}

