% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{print.tableMatrix}
\alias{print.tableMatrix}
\title{Print}
\usage{
\method{print}{tableMatrix}(x, ...)
}
\arguments{
\item{x}{\code{tableMatrix} object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{tableMatrix} method to print. Only tab data without \code{tmName} vars 
are printed.
}

