% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{[[.tableList}
\alias{[[.tableList}
\title{Double bracket}
\usage{
\method{[[}{tableList}(x, ...)
}
\arguments{
\item{x}{\code{tableList} object.}

\item{...}{Sequence of parameters for \code{tableList}.}
}
\description{
tableList method passes data.table parameters to the table attribute
}

