% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tab.tableList}
\alias{tab.tableList}
\alias{tab<-.tableList}
\title{Get or set table attribute}
\usage{
\method{tab}{tableList}(obj, ...)

\method{tab}{tableList}(obj) <- value
}
\arguments{
\item{obj}{\code{tableList} object.}

\item{...}{Passed arguments.}

\item{value}{data.table}
}
\value{
Table part.
}
\description{
\code{tableList} method to get or set table attribute.
}

