% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tab.tableMatrix}
\alias{tab.tableMatrix}
\title{Get or set table attribute}
\usage{
\method{tab}{tableMatrix}(obj, matN = NULL, addRow = FALSE, resetN = NULL)
}
\description{
tableMatrix method to get or set table attribute
}

