% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tableMatrix}
\alias{tableMatrix}
\title{Creates tableMatrix obj}
\usage{
tableMatrix(dataList, tabCol, matCol)
}
\arguments{
\item{dataList}{List of data.frames or data.tables}

\item{tabCol}{List of table attribute columns names or indicies}

\item{matCol}{List of matrix attribute columns names or indicies}
}
\value{
A tableMatrix obj
}
\description{
tableMatrix constructor, creates tableMatrix obj from a list of data.frames or data.tables
}

