% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{aid}
\alias{aid}
\alias{aid<-}
\title{S3 tableMatrix generic to get or set aid attribute}
\usage{
aid(obj, ...)

aid(obj) <- value
}
\arguments{
\item{obj}{Object.}

\item{...}{Passed arguments.}

\item{value}{User defined}
}
\description{
S3 tableMatrix generic to get or set aid attribute
}

