% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{aid.tableList}
\alias{aid.tableList}
\alias{aid<-.tableList}
\title{Get or set aid attribute}
\usage{
\method{aid}{tableList}(obj, ...)

\method{aid}{tableList}(..., obj) <- value
}
\arguments{
\item{obj}{\code{tableList} object.}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{List.}
}
\description{
\code{tableList} method to get or set aid attribute.
}

