% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{cbind.tableList}
\alias{cbind.tableList}
\alias{cbind.tableMatrix}
\title{Column binding}
\usage{
\method{cbind}{tableList}(x, y)

\method{cbind}{tableMatrix}(x, y)
}
\arguments{
\item{x}{\code{tableList} or \code{tableMatrix} object.}

\item{y}{\code{tableList} or \code{tableMatrix} object.}
}
\value{
Matrix object.
}
\description{
Column binding

Column binding of tab part of \code{tableMatrix} or \code{tableList} object.
}
