% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dataType}
\alias{dataType}
\alias{dataType<-}
\title{S3 tableMatrix generic to get or set dataType attribute}
\usage{
dataType(...)

dataType(...) <- value
}
\arguments{
\item{...}{Passed arguments.}
}
\value{
List.
}
\description{
S3 tableMatrix generic to get or set dataType attribute
}
