% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dim.tableList}
\alias{dim.tableList}
\title{Dimensions}
\usage{
\method{dim}{tableList}(x)
}
\arguments{
\item{x}{\code{tableList} object.}
}
\value{
Dimensions of the table part.
}
\description{
\code{tableList} method, returns dimensions of the table attribute.
}
