% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{dimnames.tableList}
\alias{dimnames.tableList}
\title{Dimension names}
\usage{
\method{dimnames}{tableList}(x)
}
\arguments{
\item{x}{\code{tableList} object.}
}
\value{
Column and row names of the table part.
}
\description{
\code{tableList} method, returns dimension names of the table attribute.
}

