% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{getRow.tableMatrix}
\alias{getRow.tableMatrix}
\alias{setRow.tableMatrix}
\title{Get or set row from the matrix attribute}
\usage{
\method{getRow}{tableMatrix}(obj, i = NULL, repo = NULL)

\method{setRow}{tableMatrix}(obj, value, i = NULL, repo = NULL)
}
\description{
tableMatrix method to get or set a row from the matrix attribute
}

