% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{getRowRepo.tableMatrix}
\alias{getRowRepo.tableMatrix}
\title{Get a row repo}
\usage{
\method{getRowRepo}{tableMatrix}(obj, i = NULL, repo = NULL)
}
\description{
tableMatrix method to get row repo for the matrix attribute
}

