% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\docType{data}
\name{images10By10}
\alias{images10By10}
\title{Images of distributions}
\format{A data frame with 90 rows and 103 variables:
\describe{
  \item{direction}{direction of diagonal, factor "up", "down", "both"}
  \item{dimX}{dimension x of image, in pixels}
  \item{dimY}{dimension y of image, in pixels}
  \item{pixel1}{pixel of image, in number}
  \item{pixel2}{pixel of image, in number}
  \item{pixel3}{pixel of image, in number}
  \item{pixel4}{pixel of image, in number}
  \item{pixel5}{pixel of image, in number}
  \item{pixel6}{pixel of image, in number}
  \item{pixel7}{pixel of image, in number}
  \item{pixel8}{pixel of image, in number}
  \item{pixel9}{pixel of image, in number}
  \item{pixel10}{pixel of image, in number}
  \item{pixel11}{pixel of image, in number}
  \item{pixel12}{pixel of image, in number}
  \item{pixel13}{pixel of image, in number}
  \item{pixel14}{pixel of image, in number}
  \item{pixel15}{pixel of image, in number}
  \item{pixel16}{pixel of image, in number}
  \item{pixel17}{pixel of image, in number}
  \item{pixel18}{pixel of image, in number}
  \item{pixel19}{pixel of image, in number}
  \item{pixel20}{pixel of image, in number}
  \item{pixel21}{pixel of image, in number}
  \item{pixel22}{pixel of image, in number}
  \item{pixel23}{pixel of image, in number}
  \item{pixel24}{pixel of image, in number}
  \item{pixel25}{pixel of image, in number}
  \item{pixel26}{pixel of image, in number}
  \item{pixel27}{pixel of image, in number}
  \item{pixel28}{pixel of image, in number}
  \item{pixel29}{pixel of image, in number}
  \item{pixel30}{pixel of image, in number}
  \item{pixel31}{pixel of image, in number}
  \item{pixel32}{pixel of image, in number}
  \item{pixel33}{pixel of image, in number}
  \item{pixel34}{pixel of image, in number}
  \item{pixel35}{pixel of image, in number}
  \item{pixel36}{pixel of image, in number}
  \item{pixel37}{pixel of image, in number}
  \item{pixel38}{pixel of image, in number}
  \item{pixel39}{pixel of image, in number}
  \item{pixel40}{pixel of image, in number}
  \item{pixel41}{pixel of image, in number}
  \item{pixel42}{pixel of image, in number}
  \item{pixel43}{pixel of image, in number}
  \item{pixel44}{pixel of image, in number}
  \item{pixel45}{pixel of image, in number}
  \item{pixel46}{pixel of image, in number}
  \item{pixel47}{pixel of image, in number}
  \item{pixel48}{pixel of image, in number}
  \item{pixel49}{pixel of image, in number}
  \item{pixel50}{pixel of image, in number}
  \item{pixel51}{pixel of image, in number}
  \item{pixel52}{pixel of image, in number}
  \item{pixel53}{pixel of image, in number}
  \item{pixel54}{pixel of image, in number}
  \item{pixel55}{pixel of image, in number}
  \item{pixel56}{pixel of image, in number}
  \item{pixel57}{pixel of image, in number}
  \item{pixel58}{pixel of image, in number}
  \item{pixel59}{pixel of image, in number}
  \item{pixel60}{pixel of image, in number}
  \item{pixel61}{pixel of image, in number}
  \item{pixel62}{pixel of image, in number}
  \item{pixel63}{pixel of image, in number}
  \item{pixel64}{pixel of image, in number}
  \item{pixel65}{pixel of image, in number}
  \item{pixel66}{pixel of image, in number}
  \item{pixel67}{pixel of image, in number}
  \item{pixel68}{pixel of image, in number}
  \item{pixel69}{pixel of image, in number}
  \item{pixel70}{pixel of image, in number}
  \item{pixel71}{pixel of image, in number}
  \item{pixel72}{pixel of image, in number}
  \item{pixel73}{pixel of image, in number}
  \item{pixel74}{pixel of image, in number}
  \item{pixel75}{pixel of image, in number}
  \item{pixel76}{pixel of image, in number}
  \item{pixel77}{pixel of image, in number}
  \item{pixel78}{pixel of image, in number}
  \item{pixel79}{pixel of image, in number}
  \item{pixel80}{pixel of image, in number}
  \item{pixel81}{pixel of image, in number}
  \item{pixel82}{pixel of image, in number}
  \item{pixel83}{pixel of image, in number}
  \item{pixel84}{pixel of image, in number}
  \item{pixel85}{pixel of image, in number}
  \item{pixel86}{pixel of image, in number}
  \item{pixel87}{pixel of image, in number}
  \item{pixel88}{pixel of image, in number}
  \item{pixel89}{pixel of image, in number}
  \item{pixel90}{pixel of image, in number}
  \item{pixel91}{pixel of image, in number}
  \item{pixel92}{pixel of image, in number}
  \item{pixel93}{pixel of image, in number}
  \item{pixel94}{pixel of image, in number}
  \item{pixel95}{pixel of image, in number}
  \item{pixel96}{pixel of image, in number}
  \item{pixel97}{pixel of image, in number}
  \item{pixel98}{pixel of image, in number}
  \item{pixel99}{pixel of image, in number}
  \item{pixel100}{pixel of image, in number}
}}
\usage{
images10By10
}
\description{
A dataset containing samples of distributions. Images were generated
and based on direction of diagonal they made, were split into 3 categories.
Each image is only a random sample of distribution. Average of every category
shows true shape.
}
\keyword{datasets}

