% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\docType{data}
\name{images15By15}
\alias{images15By15}
\title{Images of distributions}
\format{A data frame with 90 rows and 228 variables:
\describe{
  \item{direction}{direction of diagonal, factor "up", "down", "both"}
  \item{dimX}{dimension x of image, in pixels}
  \item{dimY}{dimension y of image, in pixels}
  \item{pixel1:pixel225}{pixel of image, in number}
}}
\usage{
images15By15
}
\description{
A dataset containing samples of distributions. Images were generated
and based on direction of diagonal they made, were split into 3 categories.
Each image is only a random sample of distribution. Average of every category
shows true shape.
}
\keyword{datasets}

