% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{is.tableMatrix}
\alias{is.tableMatrix}
\title{tableMatrix test}
\usage{
is.tableMatrix(obj)
}
\arguments{
\item{obj}{\code{tableTable} object}
}
\description{
Tests if passed object is of class \class{tableMatrix}
}

