% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{mat}
\alias{mat}
\alias{mat<-}
\title{S3 tableMatrix generic to get or set matrix attribute}
\usage{
mat(obj, ...)

mat(obj) <- value
}
\arguments{
\item{obj}{Object.}

\item{...}{Passed arguments.}

\item{value}{List of matrices.}
}
\value{
when getter list of matrices
}
\description{
S3 tableMatrix generic to get or set matrix attribute
}

