% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{mat}
\alias{mat}
\alias{mat<-}
\title{S3 tableMatrix generic to get or set matrix attribute}
\usage{
mat(...)

mat(...) <- value
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}

\item{value}{Matrix.}
}
\description{
S3 tableMatrix generic to get or set matrix attribute
}

