% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{mat.tableMatrix}
\alias{mat.tableMatrix}
\alias{mat<-.tableMatrix}
\title{Get or set matrix attribute}
\usage{
\method{mat}{tableMatrix}(obj, matN = NULL, ...)

\method{mat}{tableMatrix}(obj) <- value
}
\arguments{
\item{obj}{\code{tableMatrix} object.}

\item{matN}{Integer. Index of list in mat part of \code{tableMatrix}.
When NULL, whole mat is returned.  Default NULL.}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{Matrix.}
}
\value{
Mat part of \code{tableMatrix}
}
\description{
\code{tableMatrix} method to get or set matrix attribute. Mat consists
of list of matrices.
}
\examples{
data(images8By8)
data(images10By10)
data(images15By15)

tm <- tableMatrix(list(images8By8, images10By10), list(1:3, 1:3),
list(4:ncol(images8By8),4:ncol(images10By10)))
tm2 <- tableMatrix(images15By15, 1:3, 4:ncol(images15By15))
mat(tm)

head(mat(tm, 2))

mat(tm) <- mat(tm2) 
}

