% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{matDim}
\alias{matDim}
\alias{matDim<-}
\title{S3 tableMatrix generic to get or set matDim attribute}
\usage{
matDim(...)

matDim(...) <- value
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}

\item{value}{Matrix.}
}
\description{
S3 tableMatrix generic to get or set matDim attribute
}

