% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{matDim}
\alias{matDim}
\alias{matDim<-}
\title{S3 tableMatrix generic to get or set matDim attribute}
\usage{
matDim(obj, ...)

matDim(obj) <- value
}
\arguments{
\item{obj}{Object.}

\item{...}{Passed arguments.}

\item{value}{data.table}
}
\value{
when getter \code{data.table} with aid data
}
\description{
S3 tableMatrix generic to get or set matDim attribute
}

