% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{matDim.tableMatrix}
\alias{matDim.tableMatrix}
\alias{matDim<-.tableMatrix}
\title{Get or set matDim attribute}
\usage{
\method{matDim}{tableMatrix}(obj, matN = NULL, resetN = FALSE, ...)

\method{matDim}{tableMatrix}(obj, ...) <- value
}
\arguments{
\item{obj}{\code{tableMatrix} object.}

\item{matN}{Integer. Index of list in mat part of \code{tableMatrix}. 
Default NULL. When NULL, all rows from matDim are returned.}

\item{resetN}{Logical. When TRUE, returned matDim has matN=1. Default FALSE.}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{Matrix.}
}
\description{
\code{tableMatrix} method to get or set matDim attribute
}
\examples{
data(images8By8)
data(images10By10)
data(images15By15)

tm <- tableMatrix(list(images8By8, images10By10), list(1:3, 1:3),
list(r=c(4,ncol(images8By8)),r=c(4,ncol(images10By10))))
tm2 <- tableMatrix(images15By15, 1:3, 4:ncol(images15By15))

matDim(tm)

matDim(tm, 2)

matDim(tm) <- matDim(tm2) 
}

