% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{merge.tableMatrix}
\alias{merge.tableMatrix}
\title{Merging tableMatrix}
\usage{
\method{merge}{tableMatrix}(x, y, key, ...)
}
\arguments{
\item{x}{\code{tableMatrix} object}

\item{y}{Data.table or data.frame object.}

\item{key}{Vector of shared column names to be used as key.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Merging \code{tableMatrix} with data.table or data.frame
}

