% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{print.tableMatrix}
\alias{print.tableMatrix}
\title{Print}
\usage{
\method{print}{tableMatrix}(obj)
}
\arguments{
\item{obj}{\code{tableMatrix} object}
}
\description{
tableMatrix method prints tableMatrix obj
}

