% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{print.tableMatrix}
\alias{print.tableMatrix}
\title{Print}
\usage{
\method{print}{tableMatrix}(x, ...)
}
\arguments{
\item{x}{\code{tableMatrix} object.}

\item{...}{Passed arguments.}
}
\description{
\code{tableMatrix} method, prints table attribute without \code{tmName} columns.
}
