% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{rbind.tableMatrix}
\alias{rbind.tableMatrix}
\title{S3 method to bind tableMatrix objects by row}
\usage{
rbind.tableMatrix(..., use.names = TRUE, fill = FALSE)
}
\arguments{
\item{...}{List of \code{tableMatrix}. Objects to be bind together. Their tab
or mat part can be different.}

\item{use.names}{Just as use.names in data.table. If \code{TRUE} items are bound
by matching column names. Default \code{TRUE}.}

\item{fill}{Just as fill in data.table. Used when tab parts are different. Fills
missing columns with NAs. If \code{TRUE}, \code{user.names} has to be \code{TRUE}.
Default \code{FALSE}}
}
\value{
A tableMatrix object
}
\description{
Takes a sequence of \code{tableMatrix} arugments and combines them by row. If 
matrix parts have same dimensions, it only adds rows in tab and mat part. 
Otherwise it will add new matrix to mat part of \code{tableMatrix}.
}
\examples{
data(images8By8)
dim(images8By8)

data(images10By10)
dim(images10By10) 

image8By8TM <- tableMatrix(images8By8, list(1:3), list(4:ncol(images8By8)))
#Only first two columns are used for tab part
image8By8TM2 <- tableMatrix(images8By8, list(c(1,2)), list(4:ncol(images8By8)))
image10By10TM <- tableMatrix(images10By10, list(r=c(1,3)), list(j=c(4:ncol(images10By10))))

#Different matrix part, another matrix created in mat
tm <- rbind(image8By8TM , image10By10TM)
length(mat(tm))
tab(tm)

#Same mat and tab part
tm2 <- rbind(image8By8TM , image8By8TM)
length(mat(tm2))
tab(tm2)

#Different number/names of columns of tabs, fill=TRUE 
rbind(image8By8TM,image8By8TM2, fill=TRUE)
}

