% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{[.tableList}
\alias{[.tableList}
\title{Bracket}
\usage{
\method{[}{tableList}(x, ...)
}
\arguments{
\item{x}{\code{tableList} object.}

\item{...}{Passed arguments.}
}
\value{
\code{tableList} or vector.
}
\description{
\code{tableList} method, passes data.table bracket functionality to the table attribute.
Usage is the same as in data.table[].
}
\examples{

data(images8By8)

# Create tableList from images8By8[,1:3]
TL <- tableList(images8By8[,1:3])

# Apply data.table bracket on a tableList object
TL[direction=="both"]

}

