% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{[.tableMatrix}
\alias{[.tableMatrix}
\title{Bracket}
\usage{
\method{[}{tableMatrix}(x, ...)
}
\arguments{
\item{x}{\code{tableMatrix} object.}

\item{...}{Passed arguments.}
}
\value{
\code{tableMatrix} or vector.
}
\description{
\code{tableMatrix} method, passes data.table bracket functionality to the table attribute.
}
\examples{

data(images8By8)

# Create tableMatrix from images8By8
TM <- tableMatrix(images8By8, 1:3, 4:ncol(images8By8))

# Apply data.table bracket on a tableMatrix object
TM[direction=="both"]

}
