% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{[.tableMatrix}
\alias{[.tableMatrix}
\title{Bracket}
\usage{
\method{[}{tableMatrix}(x, ...)
}
\arguments{
\item{x}{\code{table.matrix}.}

\item{...}{Sequence of data.table parameters.}
}
\description{
S3 tableMatrix method to pass data.table parameters to the table attribute.
}

