% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{[[.tableList}
\alias{[[.tableList}
\title{Double bracket}
\usage{
\method{[[}{tableList}(x, ...)
}
\arguments{
\item{x}{\code{tableList} object.}

\item{...}{Passed arguments.}
}
\value{
vector
}
\description{
\code{tableList} method, passes double bracket functionality to the table attribute.
}

