% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tab}
\alias{tab}
\alias{tab<-}
\title{S3 tableMatrix generic to get or set table attribute}
\usage{
tab(...)

tab(...) <- value
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}

\item{value}{\code{data.table}.}
}
\description{
S3 tableMatrix generic to get or set table attribute
}

