% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tab}
\alias{tab}
\alias{tab<-}
\title{S3 tableMatrix generic to get or set table attribute}
\usage{
tab(obj, ...)

tab(obj) <- value
}
\arguments{
\item{obj}{Object.}

\item{...}{Passed arguments.}

\item{value}{data.table}
}
\description{
S3 tableMatrix generic to get or set table attribute
}

