% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tab.tableMatrix}
\alias{tab.tableMatrix}
\title{Get or set table attribute}
\usage{
\method{tab}{tableMatrix}(obj, matN = NULL, addRow = FALSE, resetN = TRUE)
}
\arguments{
\item{obj}{\code{tableMatrix} object.}

\item{matN}{Integer. Index of list in mat part of \code{tableMatrix}. Default NULL.}

\item{addRow}{Logical. When specified, column \code{tm.allRow} with row 
number is added.  Default FALSE.}

\item{resetN}{Logical. Used only when matN is specified. When FALSE
matN of returned tab won't be reseted to 1. Default TRUE.}
}
\value{
Table part of \code{tableMatrix}
}
\description{
Function to get or set table attribute of \code{tableMatrix} object.
}
\examples{
data(images8By8)
data(images10By10)
data(images10By10)

tm <- tableMatrix(list(images8By8, images10By10), list(1:3, 1:3),
list(4:ncol(images8By8),4:ncol(images10By10)))
tm2 <- tableMatrix(images15By15, 1:3, 4:ncol(images15By15))
matDim(tm)

tab(tm)

#get tab only from where matN is 1
tab(tm,2)

#tm.allRow column is added
tab(tm,2, TRUE)

#matN stays the same
tab(tm,2, TRUE, FALSE)


}

