% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tableList}
\alias{tableList}
\title{S3 class tableList object}
\usage{
tableList(tabData, aidData)
}
\arguments{
\item{tabData}{Data.frame or data.table. Descriptors of data.}

\item{aidData}{Aid List. Can be used to add more information (or structure)}
}
\value{
A tableList object
}
\description{
\code{tableList} constructor, creates tableList object from a data.frame or data.table
and list of aid data. It serves to wrap meta data and additional data structures together. It acts 
like data.table.
}
\examples{
data(images8By8)
dim(images8By8)

data(images10By10)
dim(images10By10)
images10By10AsTable <- data.table::as.data.table(images10By10)

tabPart <- images8By8[,1:3]
tableList(tabPart, NULL)


tabDataTable <- images10By10AsTable[, .(direction, dimX, dimY)]
tableList(tabDataTable, NULL)

aidPart <- list("example1", "example2")
tableList(tabPart, aidPart)
}
\seealso{
\code{\link{dim.tableList}}, \code{\link{dimnames.tableList}}
   \code{\link{rbind.tableList}}
}

