% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tableList}
\alias{tableList}
\title{Creates tableList object}
\usage{
tableList(tabData, aidData)
}
\arguments{
\item{tabData}{data.frame or data.table}

\item{aidData}{Aid list data}
}
\value{
A tableList object
}
\description{
tableList constructor, creates tableList obj from a data.frame or data.table and from a aid list
}

