% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMatrix.R
\name{tableMatrix}
\alias{tableMatrix}
\title{S3 class tableMatrix object}
\usage{
tableMatrix(dataList, tabCol, matCol)
}
\arguments{
\item{dataList}{List of data.frames or data.tables}

\item{tabCol}{List of table attribute columns names or indices}

\item{matCol}{List of matrix attribute columns names or indices}
}
\value{
A tableMatrix object
}
\description{
\code{tableMatrix} constructor, creates tableMatrix object from a list of data.frames or 
data.tables. It combines features of data.table and matrix. The result is faster access to 
data. It is useful for datasets which have following condition: the main data could be stored 
as matrix and other columns are for description.
}
\examples{
data(images8By8)
dim(images8By8)

data(images10By10)
dim(images10By10)
images10By10AsTable <- as.data.table(images10by10) 

#Use data.frame as data, first 3 columns as data descriptors, rest as part for matrix
tableMatrix(images8By8, list(r=c(1:3)), list(j=c(4:ncol(images8By8))))

#Use data.table as data, first 3 columns as data descriptors, rest as part for matrix
tableMatrix(images10By10AsTable, list(r=c(1:3)), list(j=c(4:ncol(images10By10AsTable ))))

#Use data.frame as data, first 2 columns as data descriptors with column names, 
#rest as part for matrix
tableMatrix(images8By8, list(r=c("shape","dimX","dimY")), list(j=c(4:ncol(images8By8))))

\dontrun{
tableMatrix(images8By8, list(c("shape","dimX","dimY")), list(c(4:ncol(images8By8))))	
}
}

